+++ src/Moq/Linq/MockRepository.cs
@@ -25,7 +25,18 @@ namespace Moq
 		/// <typeparam name="T">The type of the mocked object to query.</typeparam>
 		public IQueryable<T> Of<T>() where T : class
 		{
-			return CreateMockQuery<T>();
+			return this.Of<T>(this.Behavior);
+		}
+
+		/// <summary>
+		/// Access the universe of mocks of the given type, to retrieve those
+		/// that behave according to the LINQ query specification.
+		/// </summary>
+		/// <param name="behavior">Behavior of the mocks.</param>
+		/// <typeparam name="T">The type of the mocked object to query.</typeparam>
+		public IQueryable<T> Of<T>(MockBehavior behavior) where T : class
+		{
+			return this.CreateMockQuery<T>(behavior);
 		}
 
 		/// <summary>
@@ -37,7 +48,20 @@ namespace Moq
 		[SuppressMessage("Microsoft.Design", "CA1006:DoNotNestGenericTypesInMemberSignatures", Justification = "By design")]
 		public IQueryable<T> Of<T>(Expression<Func<T, bool>> specification) where T : class
 		{
-			return CreateMockQuery<T>().Where(specification);
+			return this.Of<T>(specification, this.Behavior);
+		}
+
+		/// <summary>
+		/// Access the universe of mocks of the given type, to retrieve those
+		/// that behave according to the LINQ query specification.
+		/// </summary>
+		/// <param name="specification">The predicate with the setup expressions.</param>
+		/// <param name="behavior">Behavior of the mocks.</param>
+		/// <typeparam name="T">The type of the mocked object to query.</typeparam>
+		[SuppressMessage("Microsoft.Design", "CA1006:DoNotNestGenericTypesInMemberSignatures", Justification = "By design")]
+		public IQueryable<T> Of<T>(Expression<Func<T, bool>> specification, MockBehavior behavior) where T : class
+		{
+			return this.CreateMockQuery<T>(behavior).Where(specification);
 		}
 
 		/// <summary>
@@ -48,7 +72,19 @@ namespace Moq
 		[EditorBrowsable(EditorBrowsableState.Never)]
 		public T OneOf<T>() where T : class
 		{
-			return CreateMockQuery<T>().First<T>();
+			return this.OneOf<T>(this.Behavior);
+		}
+
+		/// <summary>
+		/// Creates an mock object of the indicated type.
+		/// </summary>
+		/// <param name="behavior">Behavior of the mock.</param>
+		/// <typeparam name="T">The type of the mocked object.</typeparam>
+		/// <returns>The mocked object created.</returns>
+		[EditorBrowsable(EditorBrowsableState.Never)]
+		public T OneOf<T>(MockBehavior behavior) where T : class
+		{
+			return this.CreateMockQuery<T>(behavior).First();
 		}
 
 		/// <summary>
@@ -61,26 +97,38 @@ namespace Moq
 		[EditorBrowsable(EditorBrowsableState.Never)]
 		public T OneOf<T>(Expression<Func<T, bool>> specification) where T : class
 		{
-			return CreateMockQuery<T>().First<T>(specification);
+			return this.OneOf<T>(specification, this.Behavior);
+		}
+
+		/// <summary>
+		/// Creates an mock object of the indicated type.
+		/// </summary>
+		/// <param name="specification">The predicate with the setup expressions.</param>
+		/// <param name="behavior">Behavior of the mock.</param>
+		/// <typeparam name="T">The type of the mocked object.</typeparam>
+		/// <returns>The mocked object created.</returns>
+		[SuppressMessage("Microsoft.Design", "CA1006:DoNotNestGenericTypesInMemberSignatures", Justification = "By Design")]
+		[EditorBrowsable(EditorBrowsableState.Never)]
+		public T OneOf<T>(Expression<Func<T, bool>> specification, MockBehavior behavior) where T : class
+		{
+			return this.CreateMockQuery<T>(behavior).First(specification);
 		}
 
 		/// <summary>
 		/// Creates the mock query with the underlying queryable implementation.
 		/// </summary>
-		internal IQueryable<T> CreateMockQuery<T>() where T : class
+		internal IQueryable<T> CreateMockQuery<T>(MockBehavior behavior) where T : class
 		{
-			var method = ((Func<IQueryable<T>>)CreateQueryable<T>).GetMethodInfo();
-			return new MockQueryable<T>(Expression.Call(
-				Expression.Constant(this),
-				method));
+			var method = ((Func<MockBehavior, IQueryable<T>>)CreateQueryable<T>).GetMethodInfo();
+			return new MockQueryable<T>(Expression.Call(Expression.Constant(this), method, Expression.Constant(behavior)));
 		}
 
 		/// <summary>
 		/// Wraps the enumerator inside a queryable.
 		/// </summary>
-		internal IQueryable<T> CreateQueryable<T>() where T : class
+		internal IQueryable<T> CreateQueryable<T>(MockBehavior behavior) where T : class
 		{
-			return CreateMocks<T>().AsQueryable();
+			return this.CreateMocks<T>(behavior).AsQueryable();
 		}
 
 		/// <summary>
@@ -88,12 +136,15 @@ namespace Moq
 		/// transform the queryable query into a normal enumerable query.
 		/// This method is never used directly by consumers.
 		/// </summary>
-		private IEnumerable<T> CreateMocks<T>() where T : class
+		private IEnumerable<T> CreateMocks<T>(MockBehavior behavior) where T : class
 		{
 			do
 			{
-				var mock = this.Create<T>();
-				mock.SetupAllProperties();
+				var mock = this.Create<T>(behavior);
+				if (behavior != MockBehavior.Strict)
+				{
+					mock.SetupAllProperties();
+				}
 
 				yield return mock.Object;
 			}